# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * UCI education app API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RegistrationInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 10 */
  mobile: string;
  /** @minLength 8 */
  password: string;
}

export interface LoginBody {
  identifier: string;
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  student: 'student',
  admin: 'admin',
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  mobile: string;
  role: UserRole;
}

export interface AuthSession {
  token: string;
  expiresIn: number;
  user: User;
}

export interface UserPayload {
  user: User;
}

export type TestAnswersAnswers = {[key: string]: number | null};

export interface TestAnswers {
  answers: TestAnswersAnswers;
}

export interface GenericPayload { [key: string]: unknown }

export type GetCoursesParams = {
search?: string;
exam?: string;
};

export type GetNotesParams = {
search?: string;
subject?: string;
};

export type GetTestsParams = {
search?: string;
};