# lib/api-zod/src/generated/types/authSession.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * UCI education app API
 * OpenAPI spec version: 0.1.0
 */
import type { User } from './user';

export interface AuthSession {
  token: string;
  expiresIn: number;
  user: User;
}