# lib/api-zod/src/generated/types/registrationInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * UCI education app API
 * OpenAPI spec version: 0.1.0
 */

export interface RegistrationInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 10 */
  mobile: string;
  /** @minLength 8 */
  password: string;
}